update
[#155] correção rota /

https://jira.arinterface.com.br/work_packages/155
